Reorganize Routes and Add Edit Screen
