+++ app/(drawer)/prayergroup/[id]/_layout.tsx
@@ -15,7 +15,7 @@ const PrayerGroupLayout: React.FC = () => {
           <Stack.Screen name="edit" />
           <Stack.Screen name="about" />
           <Stack.Screen name="index" />
-          <Stack.Screen name="prayerrequest" />
+          <Stack.Screen name="prayerrequest/[id]" />
         </Stack>
       </PrayerRequestContextProvider>
     </PrayerGroupContextProvider>
