+++ app/(drawer)/prayergroup/[id]/prayerrequest/[id]/_layout.tsx
@@ -4,7 +4,8 @@ import * as React from "react";
 const PrayerRequestLayout: React.FC = () => {
   return (
     <Stack screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="[id]" />
+      <Stack.Screen name="index" />
+      <Stack.Screen name="edit" />
     </Stack>
   );
 };
