+++ app/(drawer)/prayergroup/[id]/prayerrequest/[id]/edit.tsx
@@ -1,7 +1,7 @@
 import { useLocalSearchParams } from "expo-router";
 import * as React from "react";
 
-import { PrayerRequestEditPage } from "../../../../../components/prayer-request-page/prayer-request-edit-page";
+import { PrayerRequestEditPage } from "../../../../../../components/prayer-request-page/prayer-request-edit-page";
 
 type PrayerRequestPageParams = {
   id: string;
