+++ components/navigation/drawer-constants.ts
@@ -5,4 +5,5 @@ export const ROUTES_WITHOUT_HEADER = [
   "/(drawer)/prayergroup/[id]/create",
   "/(drawer)/prayergroup/[id]/join-requests",
   "/(drawer)/prayergroup/[id]/prayerrequest/[id]",
+  "/(drawer)/prayergroup/[id]/prayerrequest/[id]/edit",
 ];
